Update based on design patterns
