+++ API/database.json
@@ -1,1 +1,1 @@
-{"A+B":{},"AA+BBA+CC":{"step":"A+C+AB","rule":"A AND A"},"A+C+AB":{"step":"A+C","rule":"A + AB"},"A+C":{},"A+B+AC":{"step":"A+B","rule":"A + AB"},"A(AB+C)B":{"step":"ADB+AD","rule":"A AND A"},"ABC+AB":{"step":"AD","rule":"A + AB"},"AB":{},"(AB)(AB)":{"step":"AB","rule":"A AND A"},"ABAB":{"step":"AB","rule":"A AND A"},"A+B+C":{}}
+{"A+B":{},"AA+BBA+CC":{"step":"A+C+AB","rule":"A AND A"},"A+C+AB":{"step":"A+C","rule":"A + AB"},"A+C":{},"A+B+AC":{"step":"A+B","rule":"A + AB"},"A(AB+C)B":{"step":"ADB+AD","rule":"A AND A"},"ABC+AB":{"step":"AD","rule":"A + AB"},"AB":{},"(AB)(AB)":{"step":"AB","rule":"A AND A"},"ABAB":{"step":"AB","rule":"A AND A"},"A+B+C":{},"AA+BB":{"step":"B+A","rule":"A AND A"},"B+A":{},"A(A+B)":{"step":"AB+A","rule":"A AND A"},"AB+A":{"step":"A","rule":"A + AB"},"A":{},"AA+BB+CC+DD":{"step":"A+B+D+C","rule":"A AND A"},"A+B+D+C":{},"AA+B+C+DD":{"step":"Z+D+H+C","rule":"A AND A"}}
